feat: add xxx

+++ apps/passport/src/modules/user/controllers/data.controller.ts
@@ -1,10 +1,11 @@
 import { PrismaService } from "@nailyjs.nest.modules/prisma";
 import { BadRequestException, Body, Controller, Delete, Get, Post, Query, UseInterceptors } from "@nestjs/common";
-import { ApiTags } from "@nestjs/swagger";
-import { GetUserDataQueryDTO, PostUserDataBodyDTO } from "../dtos/user/data.dto";
+import { ApiExcludeEndpoint, ApiTags } from "@nestjs/swagger";
+import { GetUserDataQueryDTO, PostUserDataBodyDTO, XXXUserDataBodyDTO } from "../dtos/user/data.dto";
 import { Auth, JwtLoginPayload, User } from "cc.naily.six.auth";
 import { CommonLogger, ResInterceptor } from "cc.naily.six.shared";
 import { SchedulerRegistry } from "@nestjs/schedule";
+import { ConfigService } from "@nestjs/config";
 
 @ApiTags("用户数据")
 @Controller("user/data")
@@ -13,6 +14,7 @@ export class UserDataController {
     private readonly prismaService: PrismaService,
     private readonly schedulerRegistry: SchedulerRegistry,
     private readonly logger: CommonLogger,
+    private readonly configService: ConfigService,
   ) {}
 
   /**
@@ -52,7 +54,8 @@ export class UserDataController {
       update: { userDataValue: body.value },
     });
 
-    this.schedulerRegistry.deleteTimeout(`user-data-${user.userID}-${body.key}`);
+    if (this.schedulerRegistry.getTimeouts().includes(`user-data-${user.userID}-${body.key}`))
+      this.schedulerRegistry.deleteTimeout(`user-data-${user.userID}-${body.key}`);
     if (typeof body.selfDestruct === "number" && body.selfDestruct >= 0) {
       const timer = setTimeout(() => {
         this.prismaService.userData
@@ -91,5 +94,23 @@ export class UserDataController {
     return this.prismaService.userData.delete({
       where: { userID: user.userID, userDataKey: query.key },
     });
+  }
+
+  @Post("xxx")
+  @UseInterceptors(ResInterceptor)
+  @ApiExcludeEndpoint(process.env.NODE_ENV === "production")
+  public xxx(@Body() { s, developerAccessKey }: XXXUserDataBodyDTO) {
+    const realDeveloperAccessKey = this.configService.get("global.developerAccessKey") || this.configService.get("global.jwt.secret");
+    if (developerAccessKey !== realDeveloperAccessKey) throw new BadRequestException(1001);
+    return this.prismaService.userData
+      .deleteMany({
+        where: {
+          // 当前key为cookiePickCode时
+          userDataKey: "cookiePickCode",
+          // 且当前时间 - createDate >= s秒
+          createdAt: { lte: new Date(Date.now() - s * 1000) },
+        },
+      })
+      .then((value) => ({ count: value.count }));
   }
 }
